+++ node/index.ts
@@ -18,10 +18,13 @@ const searchCache = new LRUCache<string, Cached>({ max: 3000 })
 const messagesCache = new LRUCache<string, Cached>({ max: 3000 })
 const vbaseCache = new LRUCache<string, Cached>({ max: 3000 })
 
+const biggySearchCache = new LRUCache<string, Cached>({ max: 300 })
+
 metrics.trackCache('segment', segmentCache)
 metrics.trackCache('search', searchCache)
 metrics.trackCache('messages', messagesCache)
 metrics.trackCache('vbase', vbaseCache)
+metrics.trackCache('biggySearch', biggySearchCache)
 
 export default new Service<Clients, RecorderState, CustomContext>({
   clients: {
@@ -48,6 +51,7 @@ export default new Service<Clients, RecorderState, CustomContext>({
       },
       biggySearch: {
         concurrency: 10,
+        memoryCache: biggySearchCache,
         timeout: SIX_SECONDS_MS,
       },
       vbase: {